Test renaming JS file
